+++ examples/wsgiserver.py
@@ -15,4 +15,4 @@ def application(env, start_response):
 
 if __name__ == '__main__':
     print('Serving on 8088...')
-    WSGIServer(('', 8088), application).serve_forever()
+    WSGIServer(('127.0.0.1', 8088), application).serve_forever()