+++ examples/wsgiserver_ssl.py
@@ -14,7 +14,7 @@ def hello_world(env, start_response):
     return [b'<h1>Not Found</h1>']
 
 print('Serving on https://:8443')
-server = pywsgi.WSGIServer(('', 8443), hello_world, keyfile='server.key', certfile='server.crt')
+server = pywsgi.WSGIServer(('127.0.0.1', 8443), hello_world, keyfile='server.key', certfile='server.crt')
 # to start the server asynchronously, call server.start()
 # we use blocking serve_forever() here because we have no other jobs
 server.serve_forever()