+++ src/greentest/test__refcount.py
@@ -110,7 +110,7 @@ def run_interaction(run_client):
     #s.close()
     w = weakref.ref(s._sock)
     s.close()
-    if greentest.RUNNING_ON_APPVEYOR:
+    if greentest.WIN:
         # The background thread may not have even had a chance to run
         # yet, sleep again to be sure it does. Otherwise there could be
         # strong refs to the socket still around.
@@ -137,7 +137,7 @@ def run_and_check(run_client):
         raise AssertionError('server should be dead by now')
 
 
-@greentest.skipOnAppVeyor("Often fail with timeouts; not sure why.")
+@greentest.skipOnAppVeyor("Often fail with timeouts or force closed connections; not sure why.")
 @greentest.skipOnPyPy3OnCI("Often fails with timeouts; not sure why.")
 class Test(greentest.TestCase):